Changes to package.json and Webpack to export as library
Move some packages from `devDependencies` to `dependencies` so that when
we install the package as a library we have the dependencies we need to
run the Webpack installer.

Add a NPM script `deploy:dev` and make the `postinstall` script use
Webpack in verbose mode.

Remove the "engines" section from the package.json since we want the
importing codebase to specify the engines, not this library package.

Make "react" and "react-dom" peer dependencies since we require the
importer to include those. This helps prevent us from ending up with
React version mismatches.