Modify Webpack to build and expose package as a library
Modify Webpack to build and expose package as a library. Most
importantly, we specify the Webpack config's `library` and
`libraryTarget` fields to export the code as as UMD library. We also set
Webpack `output.globalObject` to `this` so that the code can run in both
Node.js for pre-rendering and in the browser.

Reconcile some differences between the prod and development Webpack
config. I'm not really sure why these differences were in place, but
some of them were causing problems. Don't use style-loader in devel
Webpack config because it always references `window` which breaks
Node.js pre-rendering of React components. [1]

Allow the babel-loader to run on pixel-art-react for when this is
imported as a library. If we just excluded it from running on any
`node_modules` it would not process the React code inside
`pixel-art-react` when `pixel-art-react` is installed into the
importer's `node_modules` folder.

[1]: https://github.com/webpack-contrib/style-loader/issues/272