Change the index.jsx file to be library compatible
Change the index.jsx file to be library compatible. It no longer uses
ReactDOM and doesn't mount to any DOM element. I also moved the creation
of the Redux store into the created component. Creating the store on
file import doesn't seem correct when we intend for the exported
`PixelArt` component to be the entry point.